Fix denoise-loop f32 promotion, video getdata() cost, CLI transformers import
Three performance fixes from the adversarial audit: latent blends cast
back to the latent dtype so masked/warm-start loops stay bf16 (z-image
masked MLX peak 21.8->13.1 GB), video frame conversion uses the buffer
protocol (33-frame save 7.75->0.86 s, Pillow 14 safe), and transformers/
torch are off all 22 CLI entry-point import paths (~1.4->0.3 s). Training
pins its historical f32 forward pass; dtype, frame-parity, and
entry-point import contracts are test-pinned.

+++ src/mflux/models/common/latent_creator/latent_creator.py
@@ -99,4 +99,7 @@ class LatentCreator:
 
     @staticmethod
     def add_noise_by_interpolation(clean: mx.array, noise: mx.array, sigma: float) -> mx.array:
-        return (1 - sigma) * clean + sigma * noise
+        # Scheduler sigmas arrive as 0-d float32 arrays; without the cast MLX promotes bf16
+        # latents to float32 here and the entire downstream denoise loop silently runs at f32
+        # (measured 1.5-2.5x slower per step). Blend in f32, return in the latent dtype.
+        return ((1 - sigma) * clean + sigma * noise).astype(clean.dtype)
